Made a batch script to continuously run the code
When launched form the .bat file, the code runs till all the three files are created and will relaunch in case there is a code failure. To deal with system failures, the user needs to create a task using the task scheduler, which will run the code at login or startup.